+++ packages/control-property-editor-common/src/api.ts
@@ -105,7 +105,7 @@ export interface OutlineNode {
     editable: boolean;
     children: OutlineNode[];
     icon?: string;
-    extensionPointInfo?: string;
+    hasDefaultContent?: boolean;
 }
 
 export interface IconDetails {
@@ -160,6 +160,11 @@ export interface PropertyChangeDeletionDetails {
     controlId: string;
     propertyName: string;
     fileName?: string;
+}
+
+export interface ShowMessage {
+    message: string;
+    shouldHideIframe: boolean;
 }
 
 /**
@@ -226,7 +231,7 @@ export const changeProperty = createExternalAction<PropertyChange>('change-prope
 export const propertyChanged = createExternalAction<PropertyChanged>('property-changed');
 export const propertyChangeFailed = createExternalAction<PropertyChangeFailed>('change-property-failed');
 export const changeStackModified = createExternalAction<ChangeStackModified>('change-stack-modified');
-export const showMessage = createExternalAction<string>('show-dialog-message');
+export const showMessage = createExternalAction<ShowMessage>('show-dialog-message');
 export const reloadApplication = createExternalAction<void>('reload-application');
 export const storageFileChanged = createExternalAction<string>('storage-file-changed');
 export type ExternalAction =