+++ packages/create/test/__mocks__/@sap-ux/store.ts
@@ -1,8 +1,11 @@
-const mockService = {
-    read: jest.fn().mockReturnValue({})
+const store = jest.requireActual('@sap-ux/store');
+
+const mockedService = {
+    read: jest.fn().mockReturnValue(undefined)
 };
 
 module.exports = {
-    getService: jest.fn().mockResolvedValue(mockService),
-    BackendSystemKey: jest.fn()
+    ...store,
+    mockedService,
+    getService: jest.fn().mockResolvedValue(mockedService)
 };