+++ packages/preview-middleware-client/test/__mock__/sap/ui/core/ElementRegistry.ts
@@ -1,0 +1,3 @@
+export default {
+    all: jest.fn().mockReturnValue([])
+}