+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.3.6
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/project-access@1.20.0
+
+## 0.3.5
+
+### Patch Changes
+
+-   4e267684: chore - ejs upgrade
+
+## 0.3.4
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.14
+
+## 0.3.3
+
+### Patch Changes
+
+-   Updated dependencies [99bca62c]
+    -   @sap-ux/project-access@1.19.13
+
+## 0.3.2
+
+### Patch Changes
+
+-   Updated dependencies [b7d95fb3]
+    -   @sap-ux/project-access@1.19.12
+
 ## 0.3.1
 
 ### Patch Changes