+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.0.6
+
+### Patch Changes
+
+-   Updated dependencies [e3d2324c]
+    -   @sap-ux/project-access@1.20.0
+
+## 0.0.5
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.14
+
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [99bca62c]
+    -   @sap-ux/project-access@1.19.13
+
+## 0.0.3
+
+### Patch Changes
+
+-   Updated dependencies [b7d95fb3]
+    -   @sap-ux/project-access@1.19.12
+
 ## 0.0.2
 
 ### Patch Changes